- Updated Minimap redraw
git-svn-id: svn://localhost/trunk@394 a386ff7f-d321-c144-bb17-e251d28fee14

+++ KM_Defaults.pas
@@ -884,7 +884,7 @@ UnitSprite2:array[1..41,1..18]of smallint; //Sound indices vs sprite ID
   end;
 
   //Minimap tile colors, computed on tileset loading
-  TileMMColor:array[1..256]of record R,G,B:single; end;
+  TileMMColor:array[1..256]of record R,G,B:byte; end;
 
 
 implementation
